Home

IndirectImports.jl

IndirectImports.@indirect

IndirectImports.@indirect — Macro.

@indirect import Module=UUID

Define an indirectly imported Module in a downstream module.

@indirect import Module=UUID: f1, f2, ..., fn

Import functions f1, f2, ..., fn indirectly to a downstream module.

@indirect function Module.interface_function(...) ... end

Define a method of an indirectly imported function in a downstream module.

@indirect function interface_function end

Declare an interface_function in the upstream module. This function can be used and/or extended in downstream packages (via @indirect import Module=UUID) without loading the package defining interface_function. This works only at the top-level module.

Examples

Step 1: Declare indirect function in an upstream package

There must be a package that "declares" the ownership of an indirect function. Typically, such function is an interface extended by downstream packages.

To declare a function fun in a package Upstream wrap an empty definition of a function function fun end with @indirect:

module Upstream
    using IndirectImports
    @indirect function fun end
end

To define a method of an indirect function inside Upstream wrap it in @indirect:

module Upstream
    using IndirectImports
    @indirect function fun end

    @indirect fun() = 0  # defining a method
end

Step 2: Add method definition in downstream packages

First, find out the UUID of Upstream package by

julia> using Upstream

julia> Base.PkgId(Upstream)
Upstream [332e404b-d707-4859-b48f-328b8b3632c0]

Using this UUID, the Upstream package can be indirectly imported and methods for the indirect function Upstream.fun can be defined as follows:

module Downstream
    using IndirectImports
    @indirect import Upstream="332e404b-d707-4859-b48f-328b8b3632c0"
    @indirect Upstream.fun(x) = x + 1
end

Note: It looks like defining a method works without @indirect possibly due to a "bug" in Julia [1]. While it is handy to define methods without @indirect for debugging, prototyping, etc., it is a good idea to wrap the method definition in @indirect to be forward compatible with future Julia versions.

[1]
Extending a constructor is possible with only using using https://github.com/JuliaLang/julia/issues/25744

source